Move node-async-test functions to async_testing.js
I've moved most of the code from node-async-test into the
runSuitesInPaths method in async_testing.js.  This greatly
simplifies node-async-test and allows me to write a
"test/all.js" file with a single line (aside from requires):

  asyncTesting.runSuitesInPaths([path.dirname(module.filename)]);

We may want to think about combining the runSuitesInPaths utility
in with runSuites at some point.  It would be neat if runSuites
could act on dirs or filenames or modules.