A complete rewrite of node-async-testing.  Tests can now be run serially or in parallel.
The old version of node-async-testing had kind of been hacked together
quickly, and as its goals changed, the code base did its best to morph
itself to handle the new situations.  Unfortunately, this meant that at
its core, node-async-testing was written for one testing paradigm (no
suites) but was rewritten to handle another (suites).

I decided that cleaning it up was in order.  So, this is that attempt.
As to how successful I was at 'cleaning it up', I'm not sure.  However,
this does bring a major new feature!

Tests can now be run either serially or in parallel:

+ If tests are run serially (the default, and basically the same as it
  works now), then errors are caught and properly associated with their
  tests.
+ If tests are run in parallel, then they are all started at the same
  time.  This is much, much faster and very cool.  node-async-testing
  will then do its best to figure out which errors (assertions still
  work just fine) are with which tests by looking at which tests are
  still running when an error happens and then narrowing down from
  there.

At this point the rewrite isn't usable as it doesn't really report the
results of the tests. Yet.  But that will come next.  I'm working on
adding some ideas from http://github.com/caolan/nodeunit -- primarily,
the colored output and the fantastic API for building test runners.

Also, the setup and teardown functionallity has not been readded to the
rewrite. I am considering going with nodeunit's approach to that by just
making you wrap your test functions in your own setup/teardown
functions.  Simplifies things a lot!

See `test/test-everything.js` for a (very rough) example of how things
are looking.

Feedback much appreciated!

(`node-async-test`, `README.markdown` and the files in `examples`
have not been updated.)