Rename default runner and change runSuites to only run files.
Change the default runner to something more descriptive.  We'll still
know it is the default in our hearts though.

Changing runSuites to only run files/directories (and renaming it
runFiles) is a much bigger change.  If you want to run a specific
suite object, just call runSuite directly with the object.

Runners can now assume that all items are files or directories.