Deprecate `exports.wrapTests` and add better `exports.wrap`
`wrapTests` was a major pain in the butt to use.  We now have `wrap`
which is specifically for adding setup and teardown functions.

`wrap` takes an object which has three properties:

(by the way, I'm looking for a better name for this function)

suite:     a test suite object, required
setup:     a function that should be run before the test
teardown:  a function that should be run after the test

setup and teardown can both be asynchronous (and thus both must
explicitly call `done` or `finish` or whatever you want to call the
function).

    wrap({ suite: exports
         , setup: function setup(test, done) {
             test.ok(true, 'make sure we run the setup');
             test.one = extra1;
             test.two = extra2;

             done();
           }
         , teardown: function teardown(test, done) {
             test.ok(true, 'make sure we run the teardown');

             done();
           }
         });