Reorganize a bit
lib/async_testing was getting quite big. Moved testing code to
lib/testing.js. Moved running code to lib/running.js.  Left stuff that
didn't really belong in either of those where it was.

Also changed the way you register runners.  Now all information related
to a runner is kept in the exports of the module for the runner.