Add the ability for 'sub-suites'
Since suites are just objects, it is very easy to make it so you can
have 'sub suites'.  Really this is more like namespacing. All it does is
load all the tests in a sub suite into the main suite but prefixes
their name with the name of the sub suite.  Trivially easy to implement
and shouldn't slow things down at all.

I decided it was useful when writing tests for the new wrap function.
Now you can have setup/teardown functions specific to a sub suite by
'wrapping' the sub suite.

Will update the wrap tests next with some more examples.