Make the `run` method `process.exit` with the number of failed tests
Right now `run` takes three arguments:

list:      file or directory or array of files or directories
args:      array of command line arguments to change settings
callback:  a callback for when it has run all suites

This changes the third argument so that if it is not provided, when all
suites have finished this will call process.exit with a status code equal
to the number of tests that failed.  If you don't need a callback but
don't want it to do that, then pass something falsey

Now that we have this there is no reason to use the `node-async-test`
command, so it has been deprecated.