Restructer how tests are ran
This change cleans up the test running code quite a bit.  Now (more
logically) whenever a test finishes, whether by assertion failure, error
or a `test.finish()` call, it all goes through the `testFinished`
function.  Before errors were handled differently.

We also simplify the event API for writing test runners.  Before we had
4 different kind of result objects, now we have 2 (mostly).  I know this
is a backwards incompatible change, but it isn't to the public API, but
to the test runner API which I think I am the only one doing.  So, if
this is breaking your code, please let me know.

We restructred as a preface to proper setup/teardown support, which hopefully
I'll get to tonight.