Make unexpected errors kill the process
I realized I was going to A LOT of effort to catch and report on
run-time errors in code, and this was majorly complicating the way the
whole library was being written.

But I figure if there is an error, it is something that is a show
stopper anyway, and you need to be notified about it immediately.

Suite runners will be notified of the error, right before the process
exits, so they can say what happened.

If you still really want errors to be caught, use the wrap function to
add a setup to all your tests and then set the uncaughtExceptionHandler
in the test.  You'll have to run the suite in parallel then.