Re-add setup/teardown functions
Re-add some of the functionallity that I removed in
e5090414e469636702d2a7617c418726e0c58555.  Now instead of supplying a
function for a test you can supply an array of functions.  The first
and last functions in this array are considered setup and teardown
functions respectively.  If the setup function was run, then no matter
if the test passes or fails, the teardown function will be run no matter
what.  You can do this multiple times, so instead of supply an array of
length 3 you can do one of length 5.  In that case the second function
will only be run if the first one completes successfully, and the the
third the second.  However, the fourth (second to last) will be run no
matter what if the second was run and the last will be run no matter
what if the first was run.

This sounds complicated, but I don't anticipate anyone using this raw
API.  It is mainly for the wrap function to use for the suiteSetup,
suiteTeardown, setup and teardown functions.  This way you know for sure
that your teardown and suiteTeardown will be run which wasn't possible
before (suiteTeardown doesn't exist yet). This would arise if a test
failed then there was no way to make sure a teardown was run.  Which
sort of defeats the whole point of a teardown.