Don't kill process on errors that we can pin to a test
This goes back on: e5090414e469636702d2a7617c418726e0c58555 where I made
runSuite kill the process when it encountered any errors.  Now it only
does that when it encounters errors that it cannot figure out which
tests caused them.  That will only happen in parallel mode.

Why have worse results all the time when it is really only running in
parallel that causes them?