lots of work on tests,
found a couple of problems with child,

tests that logged stuff was breaking reporting,
and you could make false positives by outputting a finish message (indicating 0 tests, which is success(!))

started working on another child process implementation, using http.

but realised that the messages must be queued, so that they arrive in order.